+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 0.0.2
 =====
 
+- Add {HandBrake::Chapter#number}.
+
 0.0.1
 =====